Fix bug that caused a crash on second build
After the first build, the compiler is already initialized but the
client requests the initilization anyway to sync files. That causes that
the request (inter-isolate) is not answered, the timeout triggers and an
exception is thrown. The solution is overriding the compiler, which
causes a new check of the cache, which will detect changes in the
server side cache and is cheap, because only in memory data is
rewritten.